Extract logic to GrafanaExporterServer and make it a singleton
