+++ src/main/java/org/apache/solr/benchmarks/grafana/GrafanaExportServer.java
@@ -1,31 +1,33 @@
-package org.apache.solr.benchmarks.exporter;
+package org.apache.solr.benchmarks.grafana;
 
 import io.prometheus.client.Summary;
 import io.prometheus.client.exporter.HTTPServer;
 import org.apache.solr.benchmarks.beans.TaskType;
 import org.apache.solr.benchmarks.beans.Workflow;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
+import java.lang.invoke.MethodHandles;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-public class GrafanaExportManager {
-  private static boolean isEnabled = false;
-  private static Summary querySummary;
+public class GrafanaExportServer {
+  private final Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
+  Summary querySummary;
 
-  public static void startServer(Workflow workflow) throws IOException {
+  GrafanaExportServer(Workflow workflow) throws IOException {
     if (initMetrics(workflow)) {
+      log.info("Starting grafana exporter and metrics will be available at 127.0.0.1:{}/metrics", workflow.prometheusExportPort);
       new HTTPServer(workflow.prometheusExportPort, true);
-      isEnabled = true;
     }
   }
-
   /**
    * True if any metrics are created, hence server should be started
    * @param workflow
    * @return
    */
-  private static boolean initMetrics(Workflow workflow) {
+  final boolean initMetrics(Workflow workflow) {
     if (workflow.prometheusExportPort == null) {
       return false;
     }
@@ -44,17 +46,7 @@ public class GrafanaExportManager {
     return shouldStart;
   }
 
-  public static boolean isIsEnabled() {
-    return isEnabled;
-  }
-
-  public static void markQueryDuration(String query, long durationInNanosecond) {
-    querySummary.labels(query).observe(durationInNanosecond);
-  }
-
   private static Summary registerSummary(String name, String help, String... labelNames) {
-
     return Summary.build(name, help).labelNames(labelNames).register();
   }
-
 }
